Added two different way to create polarization maps. The results aren't the same (PROBLEM)
